pyproject.toml
[project]
name = "v0x"
version = "0.1.0"
description = "A python based api made to use telegram as a free cloud storage option"
authors = [
    {name = "EphemeralFog",email = "mohammad87115@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "cryptography (>=44.0.3,<45.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "tortoise-orm[asyncpg] (>=0.25.0,<0.26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
